fixed norm
